Add functionality to display/hide Vashj'ir
The addon would show the underwater nodes in Vashj'ir as missing, but in
reality they are marked "DISTANT" because the player was using a regular
flight node. Now, the addon will check the ID of flight master you are
talking to, and show the missing underwater nodes only if you are talking to a
seahorse flight master for use underwater. If talking to a standard
flight master, missing underwater nodes are not shown.